+++ src/OFS/ObjectManager.py
@@ -265,9 +265,7 @@ class ObjectManager(CopyContainer,
     def _subobject_permissions(self):
         import Products
         Products_permissions = getattr(Products, '__ac_permissions__', ())
-        return (Products_permissions +
-                self.aq_acquire('_getProductRegistryData')('ac_permissions')
-                )
+        return Products_permissions
 
     def filtered_meta_types(self, user=None):
         # Return a list of the types for which the user has